mpidc/pyproject.toml
[project]
name = "mpidc-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python =">=3.10,<4"
dependencies = [
    "django (==5.1.5)",
    "aiohappyeyeballs (==2.4.4)",
    "aiohttp (==3.11.11)",
    "aiohttp-retry (==2.9.1)",
    "aiosignal (==1.3.2)",
    "asgiref (==3.8.1)",
    "attrs (==25.1.0)",
    "build (==1.2.2.post1)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "django-cors-headers (==4.6.0)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.4.0)",
    "frozenlist (==1.5.0)",
    "idna (==3.10)",
    "multidict (==6.1.0)",
    "packaging (==24.2)",
    "pip-tools (==7.4.1)",
    "propcache (==0.2.1)",
    "psycopg2 (==2.9.10)",
    "psycopg2-binary (==2.9.10)",
    "pyjwt (==2.10.1)",
    "pyproject-hooks (==1.2.0)",
    "python-decouple (==3.8)",
    "requests (==2.32.3)",
    "setuptools (==75.8.0)",
    "sqlparse (==0.5.3)",
    "twilio (==9.4.4)",
    "tzdata (==2025.1)",
    "urllib3 (==2.3.0)",
    "wheel (==0.45.1)",
    "yarl (==1.18.3)",
    "reportlab (==4.3.0)",
    "chardet (==5.2.0)",
    "drf-yasg (==1.21.9)",
    "inflection (==0.5.1)",
    "pillow (==11.1.0)",
    "pytz (==2025.1)",
    "pyyaml (==6.0.2)",
    "uritemplate (==4.1.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
